backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy

openai
chromadb
tiktoken
python-multipart
gitpython
